Correct handling of Data contexts with different signatures/projections
Previous code relied on the assumption that a projection method or signature
would be present in every data context.  Fallback is to select PCA 1,2 or if
it is not there, just the first projection method.  For signatures, if the
selected signature doesn't exist anymore, selec the top sorted signature
of the selected projection.

Signed-off-by: David DeTomaso <davedeto@gmail.com>